requirements.txt
google-generativeai
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
PyPDF2
transformers
django
markdown